% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{BIC}
\alias{BIC}
\alias{BIC.flowClustList}
\alias{BIC.flowClust}
\title{Extract the BIC for a flowClust fit.}
\usage{
\method{BIC}{flowClustList}(object, ...)

\method{BIC}{flowClust}(object, ...)
}
\arguments{
\item{object}{\code{flowClustList} or \code{flowClust} fitted object}

\item{...}{other arguments. Currently not used.}
}
\value{
\code{vector} of BIC or ICL values
}
\description{
Extract the bayesian information criterion for a flowClust fit.
}
