\name{flowClust-internal}
\title{Internal flowClust functions}
\alias{.ellipsePoints}
\alias{coerce,flowClust,logical-method}
\alias{coerce,flowClust,filterResult-method}
\alias{coerce,flowClust,tmixFilterResult-method}
\alias{coerce,tmixFilterResult,logical-method}
\alias{coerce,flowClustList,flowClust-method}
\alias{coerce,flowClustList,logical-method}
\alias{coerce,flowClustList,filterResult-method}
\alias{coerce,flowClustList,tmixFilterResult-method}
\alias{coerce,tmixFilterResultList,tmixFilterResult-method}
\alias{coerce,tmixFilterResultList,logical-method}
\alias{length,tmixFilterResultList-method}
\description{
Internal functions not intended to be called directly by users.
}

\keyword{internal}
