% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{flowClust.den}
\alias{flowClust.den}
\title{generate the curve that reflects the tmixture fitting outcome}
\usage{
flowClust.den(x, obj, subset, include)
}
\arguments{
\item{x}{the numeric vector represents the x coordinates in plot}

\item{obj}{the flowClust object}

\item{subset}{An integer indicating which variable is selected for the plot.
Alternatively, a character string containing the name of the variable is
allowed if \code{x@varNames} is not \code{NULL}.}

\item{include}{A numeric vector specifying which clusters are shown on the
plot.  By default, all clusters are included.}
}
\description{
generate the curve that reflects the tmixture fitting outcome
}
