\name{show,flowClust-method}
\docType{methods}
\alias{show,flowClust-method}
\alias{show.flowClust}
\alias{show,tmixFilterResult-method}
\alias{show.tmixFilterResult}

\title{Show Method for flowClust / tmixFilterResult Object} 

\description{
This method lists out the slots contained in a \code{flowClust} object.
}

\usage{
\S4method{show}{flowClust}(object)

\S4method{show}{tmixFilterResult}(object)
}

\arguments{
\item{object}{Object returned from \code{\link{flowClust}} or \code{\link[=tmixFilter]{filter}}.}
}

\author{
Raphael Gottardo <\email{raph@stat.ubc.ca}>, Kenneth Lo <\email{c.lo@stat.ubc.ca}>
}

\seealso{
\code{\link{flowClust}}, \code{\link[=tmixFilter]{filter}}, \code{\link[=summary.flowClust]{summary}}
}

\keyword{print}





