\name{show,tmixFilter-method}
\docType{methods}
\alias{show,tmixFilter-method}
\alias{show.tmixFilter}

\title{Show Method for tmixFilter Object} 

\description{
This method shows the filtering settings stored in a \code{tmixFilter} object.
}

\usage{
\S4method{show}{tmixFilter}(object)
}

\arguments{
\item{object}{Object of class \code{\link{tmixFilter}}.}
}

\author{
Raphael Gottardo <\email{raph@stat.ubc.ca}>, Kenneth Lo <\email{c.lo@stat.ubc.ca}>
}

\keyword{print}



