\name{flowClust-internal}
\title{Internal flowClust functions}
\alias{.ellipsePoints}
\alias{coerce,flowClust,logical-method}
\alias{coerce,flowClust,filterResult-method}
\alias{coerce,flowClust,tmixFilterResult-method}
\alias{coerce,tmixFilterResult,logical-method}
\description{
Internal functions not intended to be called directly by users.
}

\keyword{internal}
