% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\docType{methods}
\name{show,flowClust-method}
\alias{show,flowClust-method}
\alias{show,flowClustList-method}
\alias{show,tmixFilter-method}
\alias{show,tmixFilterResult-method}
\alias{show,tmixFilterResultList-method}
\alias{show.flowClust}
\alias{show.tmixFilterResult}
\title{Show Method for flowClust / tmixFilterResult Object}
\usage{
\S4method{show}{flowClust}(object)

\S4method{show}{flowClustList}(object)

\S4method{show}{tmixFilter}(object)

\S4method{show}{tmixFilterResult}(object)

\S4method{show}{tmixFilterResultList}(object)
}
\arguments{
\item{object}{Object returned from \code{\link{flowClust}} or
\code{\link[=tmixFilter]{filter}}.}
}
\description{
This method lists out the slots contained in a \code{flowClust} object.
}
\author{
Raphael Gottardo <\email{raph@stat.ubc.ca}>, Kenneth Lo
<\email{c.lo@stat.ubc.ca}>
}
\seealso{
\code{\link{flowClust}}, \code{\link[=tmixFilter]{filter}},
\code{\link[=summary.flowClust]{summary}}
}
\keyword{print}

